Use DiffCallback to update the RecyclerView instead of notifyDatasetChanged() in PeopleListFragment.kt.
Also, fix bug where multiple requests for the same data from SWAPI were made. This is because the decision to make a request was dependent on an asynchronous value set by the viewmodel. Instead, we will remove the onScrollListener when a request is made and only decide whether to add it back when a new dataset is received from the viewmodel.